fix: Separate APIGW Static Hosting tests into multiple files for timing
